wormchain: split wasm handlers into multiple files and permit bank methods
