+++ wormhole_chain/app/wasm_handlers/capability.go
@@ -1,0 +1,28 @@
+package wasm_handlers
+
+import (
+	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	capabilitykeeper "github.com/cosmos/cosmos-sdk/x/capability/keeper"
+	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
+)
+
+// This defines which modules we actually want to expose cosmwasm contracts.
+// Right now we only permit methods that are read-only.
+// See https://github.com/CosmWasm/wasmd/blob/d63bea442bedf5b3055f3821472c7e6cafc3d813/x/wasm/types/expected_keepers.go
+
+type CapabilityKeeperHandler struct {
+	ScopedKeeper capabilitykeeper.ScopedKeeper
+}
+
+var _ wasmtypes.CapabilityKeeper = &CapabilityKeeperHandler{}
+
+func (b *CapabilityKeeperHandler) GetCapability(ctx sdk.Context, name string) (*capabilitytypes.Capability, bool) {
+	return b.ScopedKeeper.GetCapability(ctx, name)
+}
+func (b *CapabilityKeeperHandler) ClaimCapability(ctx sdk.Context, cap *capabilitytypes.Capability, name string) error {
+	return b.ScopedKeeper.ClaimCapability(ctx, cap, name)
+}
+func (b *CapabilityKeeperHandler) AuthenticateCapability(ctx sdk.Context, cap *capabilitytypes.Capability, name string) bool {
+	return b.ScopedKeeper.AuthenticateCapability(ctx, cap, name)
+}
